LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.get.zero.worker-expected.txt


PASS getImageData() throws INDEX_SIZE_ERR if size is zero
